refactor(parser): extract statements
